Continued on making blackjack
I continued on making blackjack. I've made the main part of the game, where the player must choose if they want to pick up another card. I also created the computer AI which decides if it wants to pick up another card. I'm pretty certain it works, but I'm not completely sure. Hopefully it works.
Also, I improved the introduction to blackjack, with the title on it's own separate line and centred. Speaking of which, I created a function which'll print the given string in the centre of the line.