Created and imported use of classes
I've created a class, Game(), which will be used to contain some variables and functions which I will be using a lot of when making the game, like getting high scores, overwriting high scores and variables containing scores. I haven't checked too much to see if there are any error but it should be mostly stable.